Adding new container "Web Development Quote" and respective styles for it.
HTML
-------------------------------------------
Adding new <article> container in order to fit in the document the "Web Development" branch.
Adding 2 new "spacer-column" classes. (They are still likely to be changed in the future) in order to fill up the empty spaces in the document.

CSS
-------------------------------------------
Adding padding-bottom to #summary-container.
Adding multiple styles to index.html previous changes.